x/dymegovernance/types/message_set_key_value_test.go
package types

import (
	"testing"

	"dymechain/testutil/sample"

	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
)

func TestMsgSetKeyValue_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgSetKeyValue
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgSetKeyValue{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgSetKeyValue{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
